+++ lib/commercetools-import/src/Models/Importcontainers/RetentionPolicyCollection.php
@@ -1,0 +1,60 @@
+<?php
+
+declare(strict_types=1);
+/**
+ * This file has been auto generated
+ * Do not change it.
+ */
+
+namespace Commercetools\Import\Models\Importcontainers;
+
+use Commercetools\Base\MapperSequence;
+use Commercetools\Exception\InvalidArgumentException;
+use stdClass;
+
+/**
+ * @template T of RetentionPolicy
+ * @extends MapperSequence<T>
+ * @psalm-method T current()
+ * @psalm-method T end()
+ * @psalm-method T at($offset)
+ * @method RetentionPolicy current()
+ * @method RetentionPolicy end()
+ * @method RetentionPolicy at($offset)
+ */
+class RetentionPolicyCollection extends MapperSequence
+{
+    /**
+     * @psalm-assert T $value
+     * @psalm-param T|stdClass $value
+     * @throws InvalidArgumentException
+     *
+     * @return RetentionPolicyCollection
+     */
+    public function add($value)
+    {
+        if (!$value instanceof RetentionPolicy) {
+            throw new InvalidArgumentException();
+        }
+        $this->store($value);
+
+        return $this;
+    }
+
+    /**
+     * @psalm-return callable(int):?T
+     */
+    protected function mapper()
+    {
+        return function (?int $index): ?RetentionPolicy {
+            $data = $this->get($index);
+            if ($data instanceof stdClass) {
+                /** @var T $data */
+                $data = RetentionPolicyModel::of($data);
+                $this->set($data, $index);
+            }
+
+            return $data;
+        };
+    }
+}